Add graph manager (add-semantics-service-DP1)
