+++ processors/optional_processors/optional_processor.rb
@@ -13,11 +13,19 @@ end
 
 post '/geo_location/keep_country' do
   puts 'Provided data are designed to be kept in their original country.'
+  puts params
+  puts params[:url_index]
+  puts params[:urls][params[:url_index].to_i]
 
   # Query node to get next nodes:
 
 
-  RestClient.post url.to_s, file_type: params[:file_type], file: data, docker_id: @docker_id, jid: jid
+  RestClient.post params[:urls][params[:url_index].to_i],
+                  file_type: params[:file_type],
+                  file: params[:file],
+                  docker_id: @docker_id,
+                  urls: params[:urls],
+                  url_index: params[:url_index].to_i + 1
 end
 
 post '/notifications/send_email' do
@@ -25,5 +33,10 @@ post '/notifications/send_email' do
 
   # Query file_type by id, to get country?
 
-  RestClient.post url.to_s, file_type: params[:file_type], file: data, docker_id: @docker_id, jid: jid
+  RestClient.post params[:urls][params[:url_index].to_i],
+                  file_type: params[:file_type],
+                  file: params[:file],
+                  docker_id: @docker_id,
+                  urls: params[:urls],
+                  url_index: params[:url_index].to_i + 1
 end
